Annotate Structural Variants

Annotate Structural Variants

Usage

annotate_sv(sv, return_simple = TRUE, return_query_params = FALSE)

Arguments

sv

a sv file in long format (similar to maf file)

return_simple

Default is TRUE where only a set of the most common columns are returned from oncoKB annotator see oncokbR::output_dictionary for more information on what is returned. If FALSE all raw columns are returned from API.

return_query_params

If TRUE, the specific parameters used to query the API are returned in new columns. This can be useful for troubleshooting the annotator. Default is FALSE.

Value

an annotated sv file

Examples

sv <- blca_sv[1:10,]

x <- annotate_sv(sv = sv)
#> Error in rename(., query_hugo_symbol = "hugo_symbol"): Can't rename columns that don't exist.
#> ✖ Column `hugo_symbol` doesn't exist.