Clean API Query Results

Clean API Query Results

Usage

.clean_query_results(
  query_result,
  original_data,
  return_simple,
  return_query_params
)

Arguments

query_result

a data frame returned from API

original_data

the original user input data frame

return_simple

Default is TRUE where only a set of the most common columns are returned from oncoKB annotator see oncokbR::output_dictionary for more information on what is returned. If FALSE all raw columns are returned from API.

return_query_params

If TRUE, the specific parameters used to query the API are returned in new columns. This can be useful for troubleshooting the annotator. Default is FALSE.

Value

a cleaned data frame