Helper to select columns for output of annotation functions

Helper to select columns for output of annotation functions

Usage

.get_select_columns(results, return_simple, return_query_params)

Arguments

results

input annotation result data frame

return_simple

Default is TRUE where only a set of the most common columns are returned from oncoKB annotator see oncokbR::output_dictionary for more information on what is returned. If FALSE all raw columns are returned from API.

return_query_params

If TRUE, the specific parameters used to query the API are returned in new columns. This can be useful for troubleshooting the annotator. Default is FALSE.

Value

a string of select column names